requirements.txt
langchain-google-genai
langgraph
pinecone-client
arxiv
tavily-python-client
pydantic
python-dotenv
google-genai
google-generativeai
pypdf